requirements.txt
websocket==0.2.1
requests==2.18.4
pyjwt==1.5.3
